requirements.txt
streamlit==1.32.0
requests==2.31.0
Pillow==10.2.0
llama-cpp-python==0.2.40
numpy==1.26.3
websocket-client==1.7.0
pytest==7.4.3
openfabric-pysdk>=0.1.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
